# Phase 5: Performance & Resource Usage optimization